평점작업 완료!

+++ semiProjectGroup1/src/controller/Movie_DeleteRateController.java
@@ -21,20 +21,27 @@ public class Movie_DeleteRateController extends HttpServlet {
 	protected void service(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
 		int movieNum = Integer.parseInt(request.getParameter("movieNum"));
-		String userId = (String) request.getSession().getAttribute("id");
+		String id = (String) request.getSession().getAttribute("id");
+		String userId = request.getParameter("userId");
+		if (userId.equals(id)) {
+			Movie_RateTableDao rateDao = Movie_RateTableDao.getInstance();
+			int n = rateDao.delete(movieNum, userId);
 
-		Movie_RateTableDao rateDao = Movie_RateTableDao.getInstance();
-		int n = rateDao.delete(movieNum, userId);
-
-		JSONObject json = new JSONObject();
-		if (n > 0) {
-			json.put("code", "success");
+			if (n > 0) {
+				request.setAttribute("movieNum", movieNum);
+				request.setAttribute("detail", "rate");
+				request.getRequestDispatcher("/Movie/review.do").forward(request, response);
+			} else {
+				request.setAttribute("movieNum", movieNum);
+				request.setAttribute("detail", "rate");
+				request.setAttribute("code", "fail");
+				request.getRequestDispatcher("/Movie/review.do").forward(request, response);
+			}
 		} else {
-			json.put("code", "fail");
+			request.setAttribute("movieNum", movieNum);
+			request.setAttribute("detail", "rate");
+			request.setAttribute("code", "fail");
+			request.getRequestDispatcher("/Movie/review.do").forward(request, response);
 		}
-		response.setContentType("text/plain;charset=utf-8");
-		PrintWriter pw = response.getWriter();
-		pw.print(json.toString());
 	}
-
 }
